backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pymongo==4.7.1
pydantic[email]==2.7.1
PyJWT==2.8.0
pdfminer.six==20221105
python-multipart==0.0.9
